requirements.txt
# Core dependencies
selenium>=4.10.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0

# Web automation
webdriver-manager>=4.0.0
undetected-chromedriver>=3.5.0

# Data processing
numpy>=1.24.0
openpyxl>=3.1.0
python-docx>=0.8.11

# Utilities
tqdm>=4.65.0
colorama>=0.4.6
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0

aiohttp==3.9.1
google-generativeai==0.3.1
asyncio==3.4.3
lxml==4.9.3
fake-useragent==1.4.0 